codes/requirements.txt
#nsml:nvcr.io/nvidia/pytorch:19.10-py3
torchvision
torchlars
torch_optimizer